+++ test/request.js
@@ -1,4 +1,4 @@
-// Copyright © 2017, 2018 IBM Corp. All rights reserved.
+// Copyright © 2017, 2021 IBM Corp. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -34,11 +34,13 @@ describe('#unit Check request response error callback', function() {
       .get('/good')
       .reply(200, { ok: true });
 
-    db.get('good', function(err) {
-      err = error.convertResponseError(err);
-      assert.strictEqual(err, undefined);
+    db.service.getDocument({ db: db.db, docId: 'good' }).then(response => {
+      assert.ok(response.result);
       assert.ok(couch.isDone());
       done();
+    }).catch(err => {
+      err = error.convertResponseError(err);
+      done(err);
     });
   });
 
@@ -46,62 +48,91 @@ describe('#unit Check request response error callback', function() {
     var couch = nock(url)
       .get('/bad')
       .times(3)
-      .reply(500, { error: 'foo', reason: 'bar' });
+      .reply(500, function(uri, requestBody) {
+        this.req.response.statusMessage = 'Internal Server Error';
+        return { error: 'foo', reason: 'bar' };
+      });
 
-    db.get('bad', function(err) {
+    db.service.getDocument({ db: db.db, docId: 'bad' }).then(response => {
+      done(new Error('Successful response when error expected.'));
+    }).catch(err => {
       err = error.convertResponseError(err);
       assert.strictEqual(err.name, 'HTTPFatalError');
-      assert.strictEqual(err.message, `500 : GET ${url}/bad - Error: foo, Reason: bar`);
+      assert.strictEqual(err.message, `500 Internal Server Error: get ${url}/bad - Error: foo, Reason: bar`);
       assert.ok(couch.isDone());
       done();
+    }).catch(err => {
+      // Handle assertion errors
+      done(err);
     });
   });
 
   it('should callback with error after 3 POST 503 responses', function(done) {
     var couch = nock(url)
-      .post('/bad')
+      .post('/_bulk_get')
       .query(true)
       .times(3)
-      .reply(503, { error: 'service_unavailable', reason: 'Service unavailable' });
-
-    db.server.request(
-      { method: 'POST', db: db.config.db, path: 'bad', qs: { revs: true }, body: {} },
-      function(err, body) {
-        assert.ok(err);
-        err = error.convertResponseError(err);
-        assert.strictEqual(err.name, 'HTTPFatalError');
-        assert.strictEqual(err.message, `503 : POST ${url}/bad - Error: service_unavailable, Reason: Service unavailable`);
-        assert.ok(couch.isDone());
-        done();
+      .reply(503, function(uri, requestBody) {
+        this.req.response.statusMessage = 'Service Unavailable';
+        return { error: 'service_unavailable', reason: 'Service unavailable' };
       });
+
+    db.service.postBulkGet({ db: db.db, revs: true, docs: [] }).then(response => {
+      done(new Error('Successful response when error expected.'));
+    }).catch(err => {
+      err = error.convertResponseError(err);
+      assert.strictEqual(err.name, 'HTTPFatalError');
+      assert.strictEqual(err.message, `503 Service Unavailable: post ${url}/_bulk_get - Error: service_unavailable, Reason: Service unavailable`);
+      assert.ok(couch.isDone());
+      done();
+    }).catch(err => {
+      // Handle assertion errors
+      done(err);
+    });
   });
 
   it('should callback with error after 3 429 responses', function(done) {
     var couch = nock(url)
       .get('/bad')
       .times(3)
-      .reply(429, { error: 'foo', reason: 'bar' });
+      .reply(429, function(uri, requestBody) {
+        this.req.response.statusMessage = 'Too Many Requests';
+        return { error: 'foo', reason: 'bar' };
+      });
 
-    db.get('bad', function(err) {
+    db.service.getDocument({ db: db.db, docId: 'bad' }).then(response => {
+      done(new Error('Successful response when error expected.'));
+    }).catch(err => {
       err = error.convertResponseError(err);
       assert.strictEqual(err.name, 'HTTPFatalError');
-      assert.strictEqual(err.message, `429 : GET ${url}/bad - Error: foo, Reason: bar`);
+      assert.strictEqual(err.message, `429 Too Many Requests: get ${url}/bad - Error: foo, Reason: bar`);
       assert.ok(couch.isDone());
       done();
+    }).catch(err => {
+      // Handle assertion errors
+      done(err);
     });
   });
 
   it('should callback with fatal error for 404 response', function(done) {
     var couch = nock(url)
       .get('/bad')
-      .reply(404, { error: 'foo', reason: 'bar' });
+      .reply(404, function(uri, requestBody) {
+        this.req.response.statusMessage = 'Not Found';
+        return { error: 'foo', reason: 'bar' };
+      });
 
-    db.get('bad', function(err) {
+    db.service.getDocument({ db: db.db, docId: 'bad' }).then(response => {
+      done(new Error('Successful response when error expected.'));
+    }).catch(err => {
       err = error.convertResponseError(err);
       assert.strictEqual(err.name, 'HTTPFatalError');
-      assert.strictEqual(err.message, `404 : GET ${url}/bad - Error: foo, Reason: bar`);
+      assert.strictEqual(err.message, `404 Not Found: get ${url}/bad - Error: foo, Reason: bar`);
       assert.ok(couch.isDone());
       done();
+    }).catch(err => {
+      // Handle assertion errors
+      done(err);
     });
   });
 
@@ -111,50 +142,65 @@ describe('#unit Check request response error callback', function() {
       .times(3)
       .replyWithError('testing badness');
 
-    db.get('bad', function(err) {
+    db.service.getDocument({ db: db.db, docId: 'bad' }).then(response => {
+      done(new Error('Successful response when error expected.'));
+    }).catch(err => {
       const err2 = error.convertResponseError(err);
       assert.strictEqual(err, err2);
       assert.ok(couch.isDone());
       done();
+    }).catch(err => {
+      // Handle assertion errors
+      done(err);
     });
   });
 
   it('should retry request if HTTP request gets timed out', function(done) {
     var couch = nock(url)
-      .post('/good')
+      .post('/_bulk_get')
       .query(true)
       .delay(1000)
-      .reply(200, { ok: true, replyNumber: 1 })
-      .post('/good')
+      .reply(200, { results: { docs: [{ id: '1', ok: { _id: '1' } }] } })
+      .post('/_bulk_get')
       .query(true)
-      .reply(200, { ok: true, replyNumber: 2 });
+      .reply(200, { results: { docs: [{ id: '1', ok: { _id: '1' } }, { id: '2', ok: { _id: '2' } }] } });
 
-    timeoutDb.server.request(
-      { method: 'POST', db: timeoutDb.config.db, path: 'good', qs: { revs: true }, body: {} },
-      function(err, body, h) {
-        assert.ok(body);
-        assert.strictEqual(body.replyNumber, 2);
-        assert.ok(couch.isDone());
-        done();
-      });
+    timeoutDb.service.postBulkGet({ db: db.db, revs: true, docs: [] }).then(response => {
+      assert.ok(response);
+      assert.ok(response.result);
+      assert.ok(response.result.results);
+      assert.ok(response.result.results.docs);
+      assert.strictEqual(response.result.results.docs.length, 2);
+      assert.ok(couch.isDone());
+      done();
+    }).catch(err => {
+      done(err);
+    });
   });
 
   it('should callback with error code ESOCKETTIMEDOUT if 3 HTTP requests gets timed out', function(done) {
+    // Increase the timeout for this test to allow for the delays
+    this.timeout(3000);
     var couch = nock(url)
-      .post('/good')
+      .post('/_bulk_get')
       .query(true)
       .delay(1000)
       .times(3)
       .reply(200, { ok: true });
 
-    timeoutDb.server.request(
-      { method: 'POST', db: timeoutDb.config.db, path: 'good', qs: { revs: true }, body: {} },
-      function(err, body, h) {
-        assert.ok(error);
-        err = error.convertResponseError(err);
-        assert.strictEqual(err.code, 'ESOCKETTIMEDOUT');
-        assert.ok(couch.isDone());
-        done();
-      });
+    timeoutDb.service.postBulkGet({ db: db.db, revs: true, docs: [] }).then(response => {
+      done(new Error('Successful response when error expected.'));
+    }).catch(err => {
+      err = error.convertResponseError(err);
+      // Note axios returns ECONNABORTED rather than ESOCKETTIMEDOUT
+      // See https://github.com/axios/axios/issues/2710 via https://github.com/axios/axios/issues/1543`
+      assert.strictEqual(err.statusText, 'ECONNABORTED');
+      assert.strictEqual(err.message, `timeout of 500ms exceeded: post ${url}/_bulk_get ECONNABORTED`);
+      assert.ok(couch.isDone());
+      done();
+    }).catch(err => {
+      // Handle assertion errors
+      done(err);
+    });
   });
 });